test: move crystal test ATTRIBUTES to conftest.py as fixture
